Rename schema.loader.physical.jdbc.handler.impl to schema.loader.physical.jdbc.handler.dialect
