+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/metadata/schema/loader/physical/jdbc/handler/dialect/MySQLDatabaseMetaDataDialectHandlerTest.java
@@ -15,12 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.metadata.schema.loader.physical.jdbc.handler;
+package org.apache.shardingsphere.infra.metadata.schema.loader.physical.jdbc.handler.dialect;
 
-import org.apache.shardingsphere.infra.database.type.DatabaseType;
-import org.apache.shardingsphere.infra.database.type.dialect.SQLServerDatabaseType;
+import org.apache.shardingsphere.infra.database.type.dialect.MySQLDatabaseType;
+import org.apache.shardingsphere.infra.metadata.schema.loader.physical.jdbc.handler.AbstractDatabaseMetaDataDialectHandlerTest;
 import org.apache.shardingsphere.sql.parser.sql.common.constant.QuoteCharacter;
-import org.junit.Before;
 import org.junit.Test;
 
 import java.sql.SQLException;
@@ -29,31 +28,25 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.when;
 
-public final class SQLServerDatabaseMetaDataDialectHandlerTest extends AbstractDatabaseMetaDataDialectHandlerTest {
-    
-    private DatabaseType sqlServerDatabaseType;
-    
-    @Before
-    public void setUp() {
-        sqlServerDatabaseType = new SQLServerDatabaseType();
-    }
+public final class MySQLDatabaseMetaDataDialectHandlerTest extends AbstractDatabaseMetaDataDialectHandlerTest {
     
     @Test
     public void assertGetSchema() throws SQLException {
         when(getConnection().getSchema()).thenReturn(DATABASE_NAME);
-        String actualSQLServerSchema = getSchema(sqlServerDatabaseType);
-        assertThat(actualSQLServerSchema, is(DATABASE_NAME));
+        String mysqlSchema = getSchema(new MySQLDatabaseType());
+        assertThat(mysqlSchema, is(DATABASE_NAME));
     }
     
     @Test
     public void assertFormatTableNamePattern() {
-        assertThat(formatTableNamePattern(sqlServerDatabaseType), is(TABLE_NAME_PATTERN));
+        String mysqlTableNamePattern = formatTableNamePattern(new MySQLDatabaseType());
+        assertThat(mysqlTableNamePattern, is(TABLE_NAME_PATTERN));
     }
     
     @Test
     public void assertGetQuoteCharacter() {
-        QuoteCharacter actualSQLServerQuoteCharacter = getQuoteCharacter(sqlServerDatabaseType);
-        assertThat(actualSQLServerQuoteCharacter.getStartDelimiter(), is("["));
-        assertThat(actualSQLServerQuoteCharacter.getEndDelimiter(), is("]"));
+        QuoteCharacter mysqlQuoteCharacter = getQuoteCharacter(new MySQLDatabaseType());
+        assertThat(mysqlQuoteCharacter.getStartDelimiter(), is("`"));
+        assertThat(mysqlQuoteCharacter.getEndDelimiter(), is("`"));
     }
 }
