+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/metadata/schema/loader/physical/jdbc/handler/dialect/SQL92DatabaseMetaDataDialectHandlerTest.java
@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.metadata.schema.loader.physical.jdbc.handler;
+package org.apache.shardingsphere.infra.metadata.schema.loader.physical.jdbc.handler.dialect;
 
-import org.apache.shardingsphere.infra.database.type.dialect.OracleDatabaseType;
+import org.apache.shardingsphere.infra.database.type.dialect.SQL92DatabaseType;
+import org.apache.shardingsphere.infra.metadata.schema.loader.physical.jdbc.handler.AbstractDatabaseMetaDataDialectHandlerTest;
 import org.apache.shardingsphere.sql.parser.sql.common.constant.QuoteCharacter;
 import org.junit.Test;
 
@@ -27,26 +28,25 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.when;
 
-public final class OracleDatabaseMetaDataDialectHandlerTest extends AbstractDatabaseMetaDataDialectHandlerTest {
+public final class SQL92DatabaseMetaDataDialectHandlerTest extends AbstractDatabaseMetaDataDialectHandlerTest {
     
     @Test
     public void assertGetSchema() throws SQLException {
-        when(getConnection().getSchema()).thenReturn(USER_NAME);
-        when(getConnection().getMetaData()).thenReturn(getDatabaseMetaData());
-        String oracleSchema = getSchema(new OracleDatabaseType());
-        assertThat(oracleSchema, is(USER_NAME));
+        when(getConnection().getSchema()).thenReturn(DATABASE_NAME);
+        String sql92Schema = getSchema(new SQL92DatabaseType());
+        assertThat(sql92Schema, is(DATABASE_NAME));
     }
     
     @Test
     public void assertFormatTableNamePattern() {
-        String oracleTableNamePattern = formatTableNamePattern(new OracleDatabaseType());
-        assertThat(oracleTableNamePattern, is(TABLE_NAME_PATTERN.toUpperCase()));
+        String sql92TableNamePattern = formatTableNamePattern(new SQL92DatabaseType());
+        assertThat(sql92TableNamePattern, is(TABLE_NAME_PATTERN));
     }
     
     @Test
     public void assertGetQuoteCharacter() {
-        QuoteCharacter oracleQuoteCharacter = getQuoteCharacter(new OracleDatabaseType());
-        assertThat(oracleQuoteCharacter.getStartDelimiter(), is("\""));
-        assertThat(oracleQuoteCharacter.getEndDelimiter(), is("\""));
+        QuoteCharacter sql92QuoteCharacter = getQuoteCharacter(new SQL92DatabaseType());
+        assertThat(sql92QuoteCharacter.getStartDelimiter(), is("\""));
+        assertThat(sql92QuoteCharacter.getEndDelimiter(), is("\""));
     }
 }
